requirements.txt
boto3==1.14.21
botocore==1.17.21
python-dotenv==0.14.0
paramiko==2.7.1
